requirements.txt
beautifulsoup4==4.10.0
Django==4.0.2
djangorestframework==3.13.1
requests==2.27.1
selenium==4.1.0
webdriver_manager==3.5.2
